requirements.txt
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-simple-history==3.7.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.52
psycopg2-binary==2.9.10
pycountry==24.6.1
pydantic==2.10.4
python-dotenv==1.0.1
pytz==2024.2
pydub==0.25.1
requests==2.32.3
django-filter==24.3
drf-flex-fields==1.0.2
gspread==6.1.0
gspread-formatting==1.2.1
beautifulsoup4==4.13.4
openpyxl==3.1.2